dialects: (stream) remove abstract stream read and write operations
